requirements.txt
fastapi==0.104.0
fastapi-elasticsearch==0.8.2
uvicorn==0.23.2
elasticsearch==8.10.1
python-dotenv==0.20.0
